Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.drawImage.animated.poster.html

<!DOCTYPE html>
<title>Canvas test: 2d.drawImage.animated.poster</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.drawImage(document.getElementById('anim-poster-gr.png'), 0, 0);
_assertPixelApprox(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 2);


});
</script>
<img src="../images/anim-poster-gr.png" id="anim-poster-gr.png" class="resource">
