LayoutTests/canvas/philip/tests/2d.drawImage.animated.poster-expected.txt

< [index] >
2d.drawImage.animated.poster
drawImage() of an APNG draws the poster frame
References: 2d.drawImage.animated.image
Actual output:
Expected output:

Passed
